calories button to calculate calories burned
get the inputs from java class to calculate calories burned, run it to the loop in Array list(userList) to check if the id match or not, if match it will get the weight from Arraylist and look for the other details to calculate calories burned. if the id is found and the entered numbers are valid then it will run through the calories burned formula inside the loop and gives the result to user when the calories button clicked.